PersonHandler getPerson(...) returns null if not found
If the provided id is not found in the list of persons, the
PersonHandler.getPerson(...) method returns a null Person object rather
than an "empty" (default) Person.

Returning an "empty" Person causes problems for the
ServerThread.login(...) method, because it the "empty" Person cannot be
cast as a User.  However, a null Person object can be casted to User,
which will also be null.  This can be cleanly handled by checking for
null.